images: allow optional override of some tools

+++ image/setup.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+BUILD_METADATA_AGENT=${BUILD_METADATA_AGENT:-yes}
+BUILD_AWS_WIRE_LENGTHS=${BUILD_AWS_WIRE_LENGTHS:-yes}
+
 set -o xtrace
 set -o pipefail
 set -o errexit
@@ -16,6 +19,11 @@ if [[ "$(zfs list -Ho name "$DATASET" 2>/dev/null)" != "$DATASET" ]]; then
 fi
 
 cd "$TOP"
+
+#
+# Clone or update clones of the tools we need:
+#
+
 if [[ ! -d image-builder ]]; then
 	git clone git@github.com:illumos/image-builder.git \
 	    image-builder
@@ -23,34 +31,48 @@ else
 	(cd image-builder && git pull --rebase)
 fi
 
-if [[ ! -d metadata-agent ]]; then
-	git clone git@github.com:illumos/metadata-agent.git \
-	    metadata-agent
-else
-	(cd metadata-agent && git pull --rebase)
+if [[ $BUILD_METADATA_AGENT == yes ]]; then
+	if [[ ! -d metadata-agent ]]; then
+		git clone git@github.com:illumos/metadata-agent.git \
+		    metadata-agent
+	else
+		(cd metadata-agent && git pull --rebase)
+	fi
 fi
 
-if [[ ! -d aws-wire-lengths ]]; then
-	git clone git@github.com:oxidecomputer/aws-wire-lengths.git \
-	    aws-wire-lengths
-else
-	(cd aws-wire-lengths && git pull --rebase)
+if [[ $BUILD_AWS_WIRE_LENGTHS == yes ]]; then
+	if [[ ! -d aws-wire-lengths ]]; then
+		git clone git@github.com:oxidecomputer/aws-wire-lengths.git \
+		    aws-wire-lengths
+	else
+		(cd aws-wire-lengths && git pull --rebase)
+	fi
 fi
+
+#
+# Build the tools:
+#
 
 (cd image-builder && cargo build --release)
-(cd metadata-agent && cargo build --release)
-(cd aws-wire-lengths && cargo build --release)
 
-for f in \
-    metadata \
-    metadata.xml \
-    userscript.sh \
-    userscript.xml; do
-	ff="$TOP/templates/files/$f"
-	rm -f "$ff"
-	if [[ $f == metadata ]]; then
-		cp "$TOP/metadata-agent/target/release/$f" "$ff"
-	else
-		cp "$TOP/metadata-agent/$f" "$ff"
-	fi
-done
+if [[ $BUILD_AWS_WIRE_LENGTHS == yes ]]; then
+	(cd aws-wire-lengths && cargo build --release)
+fi
+
+if [[ $BUILD_METADATA_AGENT == yes ]]; then
+	(cd metadata-agent && cargo build --release)
+
+	for f in \
+	    metadata \
+	    metadata.xml \
+	    userscript.sh \
+	    userscript.xml; do
+		ff="$TOP/templates/files/$f"
+		rm -f "$ff"
+		if [[ $f == metadata ]]; then
+			cp "$TOP/metadata-agent/target/release/$f" "$ff"
+		else
+			cp "$TOP/metadata-agent/$f" "$ff"
+		fi
+	done
+fi
